Processes virtual table now uses Win32 API
This commit re-architects the windows processes virtual table to make use
of the Win32 API.  This should make the table much more responsive, and
additionally fills in some of the empty column values.

+++ osquery/filesystem/fileops.h
@@ -106,7 +106,7 @@ enum SeekMode { PF_SEEK_BEGIN = 0, PF_SEEK_CURRENT, PF_SEEK_END };
 
 #ifdef WIN32
 /// Takes a Windows FILETIME object and returns seconds since epoch
-LONGLONG filetimeToUnixtime(const FILETIME& ft);
+unsigned long long filetimeToUnixtime(const FILETIME& ft);
 
 /**
  * @brief Stores information about the last Windows async request
